pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "riskbalancer"
version = "0.1.0"
description = "Portfolio risk balancing toolkit for multi-broker statements"
readme = "README.md"
authors = [{name = "RiskBalancer"}]
requires-python = ">=3.9"
dependencies = ["pyyaml>=6.0"]

[project.urls]
Homepage = "https://example.com/riskbalancer"

[project.scripts]
riskbalancer = "riskbalancer.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
